Refactor code to update data types in StridedFit

+++ src/hotspot3/fit.py
@@ -310,13 +310,13 @@ class StridedFit(BackgroundFit):
             0, original_shape[0], self.sampling_step, dtype=np.uint32
         )[::self.interpolation_step]
  
-        best_tr_with_nan = np.full_like(array, np.nan, dtype=np.float32)
+        best_tr_with_nan = np.full_like(array, np.nan, dtype=np.float16)
         best_tr_with_nan[subsampled_indices] = best_tr
 
-        best_rmsea_with_nan = np.full_like(array, np.nan, dtype=np.float32)
+        best_rmsea_with_nan = np.full_like(array, np.nan, dtype=np.float16)
         best_rmsea_with_nan[subsampled_indices] = best_rmsea
 
-        best_quantile_with_nan = np.full_like(array, np.nan, dtype=np.float32)
+        best_quantile_with_nan = np.full_like(array, np.nan, dtype=np.float16)
         best_quantile_with_nan[subsampled_indices] = best_quantile
 
         return best_tr_with_nan, best_quantile_with_nan, best_rmsea_with_nan
